完成管理员修改用户信息
