+++ app/goods/forms.py
@@ -1,10 +1,9 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField, TextAreaField, PasswordField, SelectField
-from wtforms.validators import DataRequired, Length, Optional, Email, Regexp
+from wtforms import StringField, IntegerField, TextAreaField, SelectField
+from wtforms.validators import DataRequired, Length, Optional
 from wtforms import ValidationError
-from flask_login import current_user
 
-from app.models import User, GoodsType
+from app.models import GoodsType
 
 
 class GoodsForm(FlaskForm):
@@ -34,21 +33,3 @@ class GoodsForm(FlaskForm):
         self.brand.data = goods_obj.brand
         self.quantity.data = goods_obj.quantity
         self.details.data = goods_obj.details
-
-
-class UserForm(FlaskForm):
-    username = StringField('用户名', validators=[Length(-1, 8, '用户名不超过8个字符')])
-    email = StringField('邮箱', validators=[DataRequired('请输入正确的邮箱地址'), Email('请输入正确的邮箱地址')])
-    password = PasswordField('新密码')
-    phone_number = StringField('手机号码', validators=[Length(-1, 14, '手机号码不超过14个字符'), Regexp(r'^\d*?$', message='请输入正确的手机号码')])
-    resume = TextAreaField('简介', validators=[Length(-1, 500, '简介不超过500个字符')])
-
-    def validate_username(self, field):
-        if User.query.filter(User.id != current_user.id, User.username == field.data).count() > 0:
-            raise ValidationError('该用户名已被使用')
-
-    def set_data(self):
-        self.username.data = current_user.username
-        self.email.data = current_user.email
-        self.phone_number.data = current_user.phone_number
-        self.resume.data = current_user.resume
