+++ app/templates/manage/update_user.html
@@ -5,8 +5,8 @@
     <div class="container">
         <div class="row justify-content-center">
             <div class="col col-md-8 col-xl-4 text-center">
-                <a href="{{ url_for('.index') }}" class="float-left text-dark text-decoration-none">
-                    <i class="fas fa-angle-left"></i>
+                <a href="javascript: window.history.go(-1);" class="float-left text-dark text-decoration-none">
+                    <i class="fas fa-chevron-left"></i>
                 </a>
                 个人信息
             </div>
@@ -19,10 +19,9 @@
         <div class="col col-md-8 col-xl-4">
             <form method="post">
                 {{ form.hidden_tag() }}
-                <input type="password" style="display: none;">  <!-- 防止浏览器自动填充 -->
                 <div class="form-group">
                     {{ form.username.label }}
-                    {{ form.username(class="form-control", autocomplete="off") }}
+                    {{ form.username(class="form-control") }}
                     {% for item in form.username.errors %}<small class="text-danger">{{ item }}</small><br>{% endfor %}
                 </div>
                 <div class="form-group">
@@ -36,20 +35,14 @@
                     {% for item in form.phone_number.errors %}<small class="text-danger">{{ item }}</small><br>{% endfor %}
                 </div>
                 <div class="form-group">
-                    {{ form.password.label }}
-                    {{ form.password(class="form-control", placeholder="要修改密码时请在此输入新密码", autocomplete="new-password") }}
-                    {% for item in form.password.errors %}<small class="text-danger">{{ item }}</small><br>{% endfor %}
-                </div>
-                <div class="form-group">
                     {{ form.resume.label }}
                     {{ form.resume(class="form-control", rows="5") }}
                     {% for item in form.resume.errors %}<small class="text-danger">{{ item }}</small><br>{% endfor %}
                 </div>
                 <br>
                 <button class="btn btn-block btn-info" type="submit">提交修改</button>
-                <a href="{{ url_for('.update_user') }}" class="btn btn-block btn-secondary">取消修改</a>
             </form>
-            <a href="{{ url_for('auth_bp.logout') }}" class="btn btn-outline-danger btn-block mt-5">退出账号</a>
+{#            <a href="{{ url_for('auth_bp.logout') }}" class="btn btn-outline-danger btn-block mt-5">退出账号</a>#}
         </div>
     </div>
     </div>
